requirements.txt
# FastAPI and web framework
fastapi
uvicorn[standard]
python-multipart

# Text-to-Speech and Audio (Python 3.13 compatible)
elevenlabs
openai-whisper
sounddevice
soundfile
librosa
numpy
scipy

# Audio processing alternatives (Python 3.13 compatible)
mutagen
audioread

# PyTorch for Whisper and VAD
torch
torchaudio

# NVIDIA ACE and A2F (Critical missing dependencies)
nvidia-ace
grpcio
grpcio-tools
protobuf

# HTTP requests and async
requests
aiohttp
httpx

# Data processing
pandas
pyyaml

# Development tools
python-dotenv
langchain-core
typing-extensions

# Visualization (optional)
matplotlib
networkx



protobuf-gen
PyYAML
grpclib
googleapis-common-protos
protobuf
protobuf-gen